+++ web_website/__init__.py
@@ -1,4 +1,4 @@
-# License MIT (https://opensource.org/licenses/MIT).
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 from . import models