+++ web_website/__manifest__.py
@@ -1,28 +1,27 @@
 # Copyright 2018,2020 Ivan Yelizariev <https://it-projects.info/team/yelizariev>
 # Copyright 2018 Kolushov Alexandr <https://it-projects.info/team/KolushovAlexandr>
 # Copyright 2019 Eugene Molotov <https://it-projects.info/team/em234018>
-# License MIT (https://opensource.org/licenses/MIT).
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 {
-    "name": """Website Switcher in Backend""",
+    "name": """Multi-Brand Backend""",
     "summary": """Technical module to switch Websites in Backend similarly to Company Switcher""",
     "category": "Hidden",
     # "live_test_url": "",
     "images": [],
-    "version": "13.0.3.0.6",
+    "version": "13.0.4.0.0",
     "application": False,
     "author": "IT-Projects LLC, Ivan Yelizariev",
     "support": "apps@itpp.dev",
     "website": "https://it-projects.info/team/yelizariev",
-    "license": "Other OSI approved licence",  # MIT
-    "price": 15.00,
-    "currency": "EUR",
+    "license": "LGPL-3",
+    # "price": 0.00,
+    # "currency": "EUR",
     "depends": ["web", "website", "base_setup"],
     "external_dependencies": {"python": [], "bin": []},
     "data": [
+        "views/res_users_views.xml",
         "security/security.xml",
         "security/ir.model.access.csv",
-        "views/res_config_settings_views.xml",
-        "views/res_users_views.xml",
         "views/ir_property_views.xml",
         "views/assets.xml",
     ],