+++ web_website/models/__init__.py
@@ -1,7 +1,6 @@
-# License MIT (https://opensource.org/licenses/MIT).
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 from . import website_dependent_mixin
 from . import test_website_dependent
 from . import ir_property
-from . import res_config
 from . import res_users
 from . import ir_http