+++ web_website/tests/test_website_dependent.py
@@ -1,6 +1,6 @@
 # Copyright 2018 Ivan Yelizariev <https://it-projects.info/team/yelizariev>
 # Copyright 2019 Eugene Molotov <https://it-projects.info/team/em230418>
-# License MIT (https://opensource.org/licenses/MIT).
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 import logging
 
 from odoo.tests import common, tagged